Major cleanup - Rewrite core modules

+++ src/Utils/Random.hpp
@@ -1,37 +1,33 @@
 #pragma once
 
-#include <chrono>
-#include <functional>
 #include <random>
-
-#include "../Types.hpp"
 
 namespace Dynamo::Random {
     static std::default_random_engine rng;
-    static std::uniform_real_distribution<f32> uniform(0, 1);
+    static std::uniform_real_distribution<float> uniform(0, 1);
 
     /**
      * @brief Seed the random number generator.
      *
      * @param s Seed value.
      */
-    inline void seed(i32 s) { rng.seed(s); }
+    inline void seed(long long s) { rng.seed(s); }
 
     /**
      * @brief Get a random number in the range [0, 1).
      *
-     * @return f32
+     * @return float
      */
-    inline f32 random() { return uniform(rng); }
+    inline float random() { return uniform(rng); }
 
     /**
      * @brief Get a random number in the range [min, max).
      *
      * @param min Minimum value inclusive.
      * @param max Maximum value exclusive.
-     * @return f32
+     * @return float
      */
-    inline f32 range(f32 min, f32 max) {
+    inline float range(float min, float max) {
         return min + ((max - min) * random());
     }
 } // namespace Dynamo::Random
